refactored code in meeting wed. morning
